add pretrained class to pretrain __init__ file, refactor related import
