+++ autokeras/pretrained/__init__.py
@@ -1,0 +1,3 @@
+from autokeras.pretrained.object_detector import ObjectDetector
+from autokeras.pretrained.face_detector import FaceDetector
+from autokeras.pretrained.voice_generator.voice_generator import VoiceGenerator
