+++ examples/object_detection/object_detection_example.py
@@ -1,4 +1,4 @@
-from autokeras.pretrained.object_detector import ObjectDetector
+from autokeras.pretrained import ObjectDetector
 
 if __name__ == '__main__':
     detector = ObjectDetector()
