+++ tests/pretrained/test_object_detection.py
@@ -1,4 +1,4 @@
-from autokeras.pretrained.object_detector import ObjectDetector
+from autokeras.pretrained import ObjectDetector
 from tests.common import TEST_TEMP_DIR, clean_dir
